refactor: move pipeline scripts to pipelines/ directory
- Rename fetch_members.py to pipelines/fetch_members_all.py
- Move fetch_members_no_kontoinhaber.py to pipelines/
- Move spendenquittungen.py to pipelines/
- Update entry point: fetch-members → fetch-members-all
- Update test imports for new paths
- Add pipeline documentation to README.md
