+++ src/st_andreas/pipelines/spendenquittungen.py
